feat: Enhance comments functionality with improved submission handling and UI updates
